feat: move to hook
